Use label over div
