+++ scss/components/distro.scss
@@ -11,7 +11,7 @@
         line-height: inherit;
         padding: inherit;
 
-        div {
+        label {
             text-align: center;
             overflow: hidden;
             line-height: $input-height;
